rename footer for consistency
